+++ servant-xstatic/src/XStatic/Servant.hs
@@ -1,10 +1,11 @@
 {-# LANGUAGE ImportQualifiedPost #-}
+
 module XStatic.Servant (xstaticApp) where
 
-import XStatic qualified
-import XStatic (XStaticFile)
 import Servant.API (Raw)
 import Servant.Server (ServerT, Tagged (..))
+import XStatic (XStaticFile)
+import XStatic qualified
 
 xstaticApp :: [XStaticFile] -> ServerT Raw m
 xstaticApp = Tagged . XStatic.xstaticApp