+++ xstatic-hyperscript/src/XStatic/Hyperscript.hs
@@ -4,22 +4,8 @@
 module XStatic.Hyperscript (hyperscript, hyperscriptSSE) where
 
 import Paths_xstatic_hyperscript (version)
-import XStatic (XStaticFile (..), embedFile)
+import XStatic.TH (XStaticFile, embedXStaticFile)
 
-hyperscript :: XStaticFile
-hyperscript =
-    XStaticFile
-        { name = "hyperscript.min.js"
-        , content = $(embedFile "data/_hyperscript.min.js.gz")
-        , contentType = "text/javascript; charset=UTF-8"
-        , contentVersion = version
-        }
-
-hyperscriptSSE :: XStaticFile
-hyperscriptSSE =
-    XStaticFile
-        { name = "eventsource.js"
-        , content = $(embedFile "data/eventsource.js.gz")
-        , contentType = "text/javascript; charset=UTF-8"
-        , contentVersion = version
-        }
+hyperscript, hyperscriptSSE :: XStaticFile
+hyperscript = $(embedXStaticFile "data/hyperscript.min.js.gz" version)
+hyperscriptSSE = $(embedXStaticFile "data/eventsource.js.gz" version)