+++ xstatic-remixicon/src/XStatic/Remixicon.hs
@@ -4,25 +4,11 @@
 module XStatic.Remixicon (remixicon, remixiconCss, remixiconWoff2) where
 
 import Paths_xstatic_remixicon (version)
-import XStatic (XStaticFile (..), embedFile)
+import XStatic.TH (XStaticFile, embedXStaticFile)
 
 remixicon :: [XStaticFile]
 remixicon = [remixiconCss, remixiconWoff2]
 
-remixiconCss :: XStaticFile
-remixiconCss =
-    XStaticFile
-        { name = "remixicon.min.css"
-        , content = $(embedFile "data/remixicon.min.css.gz")
-        , contentType = "text/css"
-        , contentVersion = version
-        }
-
-remixiconWoff2 :: XStaticFile
-remixiconWoff2 =
-    XStaticFile
-        { name = "remixicon.woff2"
-        , content = $(embedFile "data/remixicon.woff2.gz")
-        , contentType = "application/font-woff2"
-        , contentVersion = version
-        }
+remixiconCss, remixiconWoff2 :: XStaticFile
+remixiconCss = $(embedXStaticFile "data/remixicon.min.css.gz" version)
+remixiconWoff2 = $(embedXStaticFile "data/remixicon.woff2.gz" version)