+++ xstatic-sakura/src/XStatic/Sakura.hs
@@ -4,32 +4,9 @@
 module XStatic.Sakura (sakuraCss, sakuraDarkSolarizedCss, normalizeCss) where
 
 import Paths_xstatic_sakura (version)
-import XStatic (XStaticFile (..), embedFile)
+import XStatic.TH (XStaticFile, embedXStaticFile)
 
-sakuraCss :: XStaticFile
-sakuraCss =
-    XStaticFile
-        { name = "sakura.css"
-        , content = $(embedFile "data/sakura.css.gz")
-        , contentType = "text/css"
-        , contentVersion = version
-        }
-
-sakuraDarkSolarizedCss :: XStaticFile
-sakuraDarkSolarizedCss =
-    XStaticFile
-        { name = "sakura-dark-solarized.css"
-        , content = $(embedFile "data/sakura-dark-solarized.css.gz")
-        , contentType = "text/css"
-        , contentVersion = version
-        }
-
-
-normalizeCss :: XStaticFile
-normalizeCss =
-    XStaticFile
-        { name = "normalize.css"
-        , content = $(embedFile "data/normalize.css.gz")
-        , contentType = "text/css"
-        , contentVersion = version
-        }
+sakuraCss, sakuraDarkSolarizedCss, normalizeCss :: XStaticFile
+sakuraCss = $(embedXStaticFile "data/sakura.css.gz" version)
+sakuraDarkSolarizedCss = $(embedXStaticFile "data/sakura-dark-solarized.css.gz" version)
+normalizeCss = $(embedXStaticFile "data/normalize.css.gz" version)