+++ xstatic-sortable/src/XStatic/Sortable.hs
@@ -4,13 +4,7 @@
 module XStatic.Sortable (sortable) where
 
 import Paths_xstatic_sortable (version)
-import XStatic (XStaticFile (..), embedFile)
+import XStatic.TH (XStaticFile, embedXStaticFile)
 
 sortable :: XStaticFile
-sortable =
-    XStaticFile
-        { name = "Sortable.min.js"
-        , content = $(embedFile "data/Sortable.min.js.gz")
-        , contentType = "text/javascript; charset=UTF-8"
-        , contentVersion = version
-        }
+sortable = $(embedXStaticFile "data/Sortable.min.js.gz" version)