+++ xstatic-tailwind/src/XStatic/Tailwind.hs
@@ -4,13 +4,7 @@
 module XStatic.Tailwind (tailwind) where
 
 import Paths_xstatic_tailwind (version)
-import XStatic (XStaticFile (..), embedFile)
+import XStatic.TH (XStaticFile, embedXStaticFile)
 
 tailwind :: XStaticFile
-tailwind =
-    XStaticFile
-        { name = "tailwind.min.js"
-        , content = $(embedFile "data/tailwind.min.js.gz")
-        , contentType = "text/javascript; charset=UTF-8"
-        , contentVersion = version
-        }
+tailwind = $(embedXStaticFile "data/tailwind.min.js.gz" version)