+++ xstatic/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add xstaticMiddleware
 - Enable StrictData
+- Remove file-embed dependency, use xstatic-th instead.
 
 ## 0.1.0